This refactors the ScaleDenominator to be a type instead of a class
