--- src/Model/ScaleDenominator/ScaleDenominator.spec.ts
@@ -1,5 +1,0 @@
-import ScaleDenominator from './ScaleDenominator';
-
-it('is defined', () => {
-  expect(ScaleDenominator).toBeDefined();
-});
