--- src/Model/ScaleDenominator/ScaleDenominator.ts
@@ -1,10 +1,0 @@
-/**
- *
- *
- * @class ScaleDenominator
- */
-class ScaleDenominator {
-
-}
-
-export default ScaleDenominator;
